Allow the plugin to issue content type headers for S3
PDFs were getting a text/html content type header when served
via an S3 stream. There's probably something more we should do
when working with S3 uploads, but the immediate solution is to
give control back to the plugin. This *may* break things again
with PDFs in IE, but we'll see.